Move ConstantPool out of ClassfileParser
Even though it's intimately related to the class file parser, it's a
large enough piece of behaviour, which is already a distinct class, that
it's reasonable to think about dealing with as a standalone "thing".
